Service account for worker controller and a few small bug fixes

+++ kubernetes/scripts/k8s/upload_piwind_model_data.sh
@@ -93,7 +93,9 @@ done
 
 for file in $OPTIONAL_MODEL_FILES; do
   file=${PWP}/$file
-  kubectl cp "${file}" host-data-volume-pod:/mnt/host/model-data/piwind/
+  if [ -f "$file" ] && ! [ -d "$file" ]; then
+    kubectl cp "${file}" host-data-volume-pod:/mnt/host/model-data/piwind/
+  fi
   echo -n .
 done
